Devide problems
Devide by 0 is now throwing exception.
Two symbols following each other are replaced by one(+- > -) or (*- > * (-digit))

+++ Classes/Abstracts/CalculatorFactory.cs
@@ -80,7 +80,7 @@ namespace SimpleCalculator.Classes.Abstracts
                     mathSymbol = new DivisionSymbol();
                     break;
                 default:
-                    mathSymbol = new DigitSymbol(int.Parse(symbol));
+                    mathSymbol = new DigitSymbol(double.Parse(symbol));
                     break;
             }
             return mathSymbol;
